+++ src/browser-UnicodeIdentifiers.js
@@ -1,0 +1,18 @@
+/*global shimIndexedDB*/
+import * as UnicodeIdentifiers from './UnicodeIdentifiers';
+
+// BEGIN: Same code as in browser.js
+import setGlobalVars from './setGlobalVars.js';
+import CFG from './cfg.js';
+
+CFG.win = typeof window !== 'undefined' ? window : self; // For Web Workers
+
+setGlobalVars();
+// END: Same code as in browser.js
+
+const __setUnicodeIdentifiers = shimIndexedDB.__setUnicodeIdentifiers.bind(shimIndexedDB);
+shimIndexedDB.__setUnicodeIdentifiers = function () {
+    __setUnicodeIdentifiers(UnicodeIdentifiers);
+};
+
+shimIndexedDB.__setUnicodeIdentifiers();